Changes to remove the new <span> tags
epguides.com changed their output to include new <span> tags that link
to various other websites, most of them recaps or trailers. That output
is useless to us and thus has to be discarded.

I've also added a license to the top of the file so that you can freely
use it as you please.